Update 11000bee.cpp
 I changed int to long long

+++ 11000bee.cpp
@@ -10,21 +10,21 @@ int main(){
 
     cin >> endyear;
 
-    int female=0;
-    int male=0;
-    int immortal=1;
+    long long female=0;
+    long long male=0;
+    long long immortal=1;
 
     if(endyear==-1) return 0;
 
     for (int i=0; i<endyear; i++){
 
-        int femaleNextyear=male;
-        int maleNextyear=female+male+immortal;
+        long long femaleNextyear=male;
+        long long maleNextyear=female+male+immortal;
 
         female=femaleNextyear;
         male=maleNextyear;
     }
-    int total=male+female+immortal;
+    long long total=male+female+immortal;
     cout << male << " " << total << endl;
     }
     while (true);
